Missing value handling in estimation and filtering and removal of parallel (using futures package now)
- Replaced all snow and doParallel functionality with future package. The cores
argument is now removed from options of all functions and instead the user
can just setup a plan via the future package which is much simpler and allows
much more customization. This may break backwards compatibility with versions
less than this which made use of the cores argument.
- Missing values are now allowed in the data for estimation. The series with
missing value are replaced with their prediction step (zero error). The
covariance and likelihood is calculated based on the joint rows
with non-missing values. The tsetsad (auto diff) is also updated to allow for
this case.
- tsaggregate and tsbacktest not fully tested/fixed yet [on next update]
- tsdecompose needs a simplify argument to align it with the other packages
method